Refactor: 컬렉션 아이템 open 상태 관리 함수를 커스텀 훅으로 변경
