+++ src/hooks/index.js
@@ -1,3 +1,4 @@
 export { default as useColorScheme } from './useColorScheme';
 export { default as useObserver } from './useObserver';
 export { default as usePagination } from './usePagination';
+export { default as useSelectedItem } from './useSelectedItem';
