+++ src/hooks/useSelectedItem.js
@@ -1,0 +1,31 @@
+import { useEffect } from 'react';
+import { useMediaQuery } from '@mantine/hooks';
+import { useRecoilValue } from 'recoil';
+import { sideNavOpenedState, categoryState } from '../recoil/atom';
+
+const useSelectedItem = (setSelectedItem, setItemSelected, selectedItem) => {
+  const middleScreen = useMediaQuery('(max-width: 51.25rem)');
+
+  const category = useRecoilValue(categoryState);
+  const isNavOpened = useRecoilValue(sideNavOpenedState);
+
+  useEffect(() => {
+    setSelectedItem(null);
+  }, [category]);
+
+  useEffect(() => {
+    if (middleScreen && isNavOpened) {
+      setItemSelected(false);
+    }
+  }, [middleScreen]);
+
+  useEffect(() => {
+    if (!middleScreen && selectedItem && isNavOpened) {
+      setItemSelected(true);
+    }
+  }, [middleScreen]);
+
+  return middleScreen;
+};
+
+export default useSelectedItem;
